organization and event creation support

+++ tigerevents/events/routes.py
@@ -1,8 +1,8 @@
 from flask import render_template, url_for, flash, redirect, request, abort, Blueprint
-from flask_login import login_user, current_user
+from flask_login import current_user
 from tigerevents import db
-from tigerevents.models import Event, User, Saved
-from tigerevents.events.utils import login_required
+from tigerevents.models import Event, Saved
+from tigerevents.utils import login_required
 from tigerevents.events.forms import EventForm
 
 
@@ -14,7 +14,7 @@ def event(event_id):
     return render_template("event.html", title=event.title, event=event)
 
 @events.route("/event/save/<int:event_id>")
-@login_required(role='ANY')
+@login_required(role='User')
 def save(event_id):
     event = Event.query.get_or_404(event_id)
 
@@ -52,7 +52,7 @@ def save(event_id):
     
 
 @events.route("/event/rsvp/<int:event_id>")
-@login_required(role='ANY')
+@login_required(role='User')
 def rsvp(event_id):
     event = Event.query.get_or_404(event_id)
 
@@ -68,14 +68,14 @@ def rsvp(event_id):
             message="Event already added.", 
             category="danger"
         )
-        return redirect(url_for("users.myevents"))
+        return redirect(url_for("main.home"))
 
     elif current_user.clash(event):
         flash(
             message="Clashes with an event in your calendar.", 
             category="danger"
         )
-        return redirect(url_for("users.myevents"))
+        return redirect(url_for("main.home"))
 
     elif current_user.is_saved(event):
         Saved.query.filter_by(event_id = event_id).filter_by(user_id = current_user.id)[0].going = True
@@ -100,7 +100,7 @@ def rsvp(event_id):
     return redirect(url_for("main.home"))
 
 @events.route("/event/remove/<int:event_id>")
-@login_required(role='ANY')
+@login_required(role='User')
 def remove(event_id):
     event = Event.query.get_or_404(event_id)
 
@@ -125,7 +125,7 @@ def remove(event_id):
 
 
 @events.route("/event/delete/<int:event_id>", methods=["POST"])
-@login_required(role="Host")
+@login_required(role="ANY")
 def delete_event(event_id):
     event = Event.query.get_or_404(event_id)
 
@@ -136,47 +136,74 @@ def delete_event(event_id):
         )
         return redirect(url_for("main.home"))
 
-    db.session.delete(event)
-    db.session.commit()
+    if event.creator == current_user:
+        db.session.delete(event)
+        db.session.commit()
 
-    flash(
-        message=f"Successfully deleted {event.title}",
-        category="success",
-    )
+        flash(
+            message=f"Successfully deleted {event.title}",
+            category="success",
+        )
+    else:
+        flash(
+            message="No authorization.", 
+            category="danger"
+        )
 
     return redirect(url_for("users.myevents"))
 
 
 @events.route("/event/new", methods=["GET", "POST"])
-@login_required(role="Host")
-def new_event(event_id):
+@login_required(role="ANY")
+def new_event():
     form = EventForm()
 
     if form.validate_on_submit():
-        event = Event(
+        new_event = Event(
             title=form.title.data,
             description=form.description.data,
             location=form.location.data,
             start_date=form.start_date.data,
             end_date=form.end_date.data,
-            host=current_user.host_of(),
+            host=current_user.org,
+            creator=current_user
         )
-        db.session.add(event)
+        db.session.add(new_event)
         db.session.commit()
 
         flash("Your event has been created!", "success")
-        return redirect(url_for("main.home"))
-    return render_template(
-        "create_event.html", title="New Event", form=form, legend="New Event"
-    )
+        return redirect(url_for("users.myevents"))
+    return render_template('create_event.html', title='New Event',
+                           form=form, legend='New Event')
 
-    flash(
-        message=f"Successfully deleted {event.title}",
-        category="success",
-    )
+@events.route("/event/edit/<int:event_id>", methods=["GET", "POST"])
+@login_required(role="ANY")
+def edit_event(event_id):
+    form = EventForm()
+    event = Event.query.filter_by(id=event_id)
 
-    return redirect(url_for("users.myevents"))
-
-
-
+    if form.validate_on_submit():
     
+        event.title = form.title.data
+        event.description = form.description.data
+        event.location = form.location.data
+        event.start_date = form.start_date.data
+        event.end_date = form.end_date.data
+        
+        db.session.commit()
+        flash("Your event has been updated!", "success")
+        if current_user.urole == 'User':
+            return redirect(url_for("users.myevents"))
+        else:
+            return redirect(url_for("main.home"))
+
+    # elif request.method == 'GET':
+
+    #     form.title.data = event.title
+    #     form.description.data = event.description
+    #     form.location.data = event.location
+    #     form.start_date.data = event.start_date
+    #     form.end_date.data = event.end_date
+        
+    return render_template('create_event.html', title='Edit Event',
+                           form=form, legend='Edit Event')
